feat(dishes): add function to add dishes in liked.

+++ src/modules/users/users-controller.ts
@@ -17,13 +17,56 @@ export class UsersController extends BasicController {
                 method: 'delete',
                 function: this.deleteUser,
             },
+            {
+                path: '/assign',
+                method: 'put',
+                function: this.assignAdmin,
+            },
         ]);
     }
 
     async deleteUser(req: Request, res: Response, next: NextFunction) {
         const body = req.body;
         checkCorrectnessOfBody(body, ['email']);
+        const { email } = body;
+
+        if (!req.user.isAdmin && req.user.email !== email) {
+            res.status(403).send();
+            return;
+        }
+
         const result = await this.usersService.deleteUser(body.email);
-        res.send(result);
+        if (!result) {
+            throw new Error("User doesn't exist.");
+        }
+        res.status(200).send('User deleted successfully.');
+    }
+
+    async assignAdmin(req: Request, res: Response, next: NextFunction) {
+        const email = req.user.email;
+        const isAdmin = req.user.isAdmin;
+        const emailForChange = req.body.email;
+
+        if (!emailForChange) {
+            res.status(400).send("User's email is not transferred.");
+            return;
+        }
+        if (!email) {
+            res.status(401).send('User is not authorized.');
+            return;
+        }
+        if (!isAdmin) {
+            res.status(403).send('Only admins can appoint admins.');
+            return;
+        }
+
+        const result = await this.usersService.assignAdmin(email);
+        if (!result) {
+            res.status(400).send(
+                `User with email ${emailForChange} doesn't exist.`,
+            );
+            return;
+        }
+        res.status(201).send(`User with email ${email} is admin now.`);
     }
 }
